+++ src/storybook/index.ts
@@ -4,6 +4,8 @@ export { patchStorybookFocusForDocs } from "./patch-storybook-focus";
 
 export { isVrCaptureStorybook, vrStorybookAddons, VR_CAPTURE_INCOMPATIBLE_ADDONS } from "./vr-storybook-addons";
 
+export { applyVrDockerViteWatch } from "./vr-vite-docker-watch";
+
 export {
   BURST_VR_TAG,
   FORCE_VR_TAG,